requirements.txt
aioflask
psycopg2-binary
flask-sqlalchemy
python-dotenv
openai
gtts
python-telegram-bot
rasa